+++ rand_core/src/impls.rs
@@ -179,19 +179,19 @@ mod test {
     fn test_fill_via_u32_chunks() {
         let src_orig = [1u32, 2, 3];
 
-        let mut src = src_orig;
+        let src = src_orig;
         let mut dst = [0u8; 11];
-        assert_eq!(fill_via_chunks(&mut src, &mut dst), (3, 11));
+        assert_eq!(fill_via_chunks(&src, &mut dst), (3, 11));
         assert_eq!(dst, [1, 0, 0, 0, 2, 0, 0, 0, 3, 0, 0]);
 
-        let mut src = src_orig;
+        let src = src_orig;
         let mut dst = [0u8; 13];
-        assert_eq!(fill_via_chunks(&mut src, &mut dst), (3, 12));
+        assert_eq!(fill_via_chunks(&src, &mut dst), (3, 12));
         assert_eq!(dst, [1, 0, 0, 0, 2, 0, 0, 0, 3, 0, 0, 0, 0]);
 
-        let mut src = src_orig;
+        let src = src_orig;
         let mut dst = [0u8; 5];
-        assert_eq!(fill_via_chunks(&mut src, &mut dst), (2, 5));
+        assert_eq!(fill_via_chunks(&src, &mut dst), (2, 5));
         assert_eq!(dst, [1, 0, 0, 0, 2]);
     }
 
@@ -199,19 +199,19 @@ mod test {
     fn test_fill_via_u64_chunks() {
         let src_orig = [1u64, 2];
 
-        let mut src = src_orig;
+        let src = src_orig;
         let mut dst = [0u8; 11];
-        assert_eq!(fill_via_chunks(&mut src, &mut dst), (2, 11));
+        assert_eq!(fill_via_chunks(&src, &mut dst), (2, 11));
         assert_eq!(dst, [1, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0]);
 
-        let mut src = src_orig;
+        let src = src_orig;
         let mut dst = [0u8; 17];
-        assert_eq!(fill_via_chunks(&mut src, &mut dst), (2, 16));
+        assert_eq!(fill_via_chunks(&src, &mut dst), (2, 16));
         assert_eq!(dst, [1, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0, 0]);
 
-        let mut src = src_orig;
+        let src = src_orig;
         let mut dst = [0u8; 5];
-        assert_eq!(fill_via_chunks(&mut src, &mut dst), (1, 5));
+        assert_eq!(fill_via_chunks(&src, &mut dst), (1, 5));
         assert_eq!(dst, [1, 0, 0, 0, 0]);
     }
 }